Allow get_block_headers_range to take negative relative values
Negatives mean "relative to the current height" and so can be used to
request the last N blocks.

+++ src/rpc/core_rpc_server_commands_defs.h
@@ -1059,8 +1059,11 @@ struct GET_CONNECTIONS : NO_ARGS {
 ///
 /// Inputs:
 ///
-/// - `start_height` -- The starting block's height.
+/// - `start_height` -- The starting block's height.  If negative then the value means relative to
+///   the current chain height (e.g. "start_height": -10, "end_height": -1 would request the last 10
+///   blocks).
 /// - `end_height` -- The ending block's height (inclusive).  Must be less than start_height + 1000.
+///   Negative values are relative to the current chain height.
 /// - `fill_pow_hash` -- Tell the daemon if it should fill out pow_hash field.
 /// - `get_tx_hashes` -- If true (default false) then include the hashes of non-coinbase
 ///   transactions
@@ -1083,11 +1086,11 @@ struct GET_BLOCK_HEADERS_RANGE : PUBLIC {
     }
 
     // Used for this endpoint as well as the by_hash/by_height versions.
-    static constexpr size_t MAX_COUNT = 1000;
+    static constexpr int64_t MAX_COUNT = 1000;
 
     struct request_parameters {
-        uint64_t start_height;
-        uint64_t end_height;
+        int64_t start_height;
+        int64_t end_height;
         bool fill_pow_hash;
         bool get_tx_hashes;
     } request;
